[CL-2101] Moved group_policy patch to core
